safari-600.7-branch/LayoutTests/platform/mac/fast/canvas/canvas-fillRect-gradient-shadow-expected.txt

Ensure correct behavior of canvas with fillRect using a gradient fillStyle and a shadow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 254.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 254.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
PASS successfullyParsed is true

TEST COMPLETE
